backend-main/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
astroid==3.2.4
certifi==2024.7.4
cfgv==3.4.0
click==8.1.7
dill==0.3.8
distlib==0.3.8
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.112.1
fastapi-cli==0.0.5
filelock==3.15.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
identify==2.6.0
idna==3.8
importlib_metadata==8.4.0
isort==5.13.2
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
nodeenv==1.9.1
numpy==2.1.0
pandas==2.2.2
platformdirs==4.2.2
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
tomli==2.0.1
tomlkit==0.13.2
typer==0.12.4
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
uvloop==0.20.0
virtualenv==20.26.3
watchfiles==0.23.0
websockets==13.0
yapf==0.40.2
zipp==3.20.0
